+++ utils/design-patterns/result-pattern.go
@@ -1,0 +1,3 @@
+package design_patterns
+
+// Way of writing error tole-rant code and handling different kinds of outcomes, such as success, failure, or exception.